+++ R/release.r
@@ -79,7 +79,7 @@ release <- function(pkg = ".", check = TRUE, args = NULL) {
       return(invisible())
   }
 
-  if (has_src(pkg)) {
+  if (pkgbuild::pkg_has_src(pkg$path)) {
     if (yesno("Have you run R CMD check with valgrind?"))
       return(invisible())
   }
@@ -262,7 +262,7 @@ submit_cran <- function(pkg = ".", args = NULL) {
 
 build_cran <- function(pkg, args) {
   message("Building")
-  built_path <- build(pkg, tempdir(), manual = TRUE, args = args)
+  built_path <- pkgbuild::build(pkg$path, tempdir(), manual = TRUE, args = args)
   message("Submitting file: ", built_path)
   message("File size: ",
           format(as.object_size(file.info(built_path)$size), units = "auto"))