+++ tests/has-devel.R
@@ -1,3 +1,0 @@
-library(devtools)
-
-has_devel()